Added motor controls.
It turns out that controlling the motor via USB/Bluetooth sucks.

You don't have the really cool fine-control that you can get via
a program running on the brick itself.

So I simplified the interface for the moment to just run and stop.

I'm looking at possible solutions (such an running a program on the
brick we can talk to).